requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
matplotlib>=3.4.3
opencv-python>=4.5.3
Pillow>=8.3.1
tqdm>=4.62.3

# Deep learning utilities
einops>=0.3.2
timm>=0.4.12  # For dataset utilities

# Development tools
jupyter>=1.0.0  # For notebooks
ipykernel>=6.4.0  # For Jupyter kernel

# Optional: for data science and visualization
pandas>=1.3.3
seaborn>=0.11.2

# For documentation
sphinx>=4.2.0
sphinx-rtd-theme>=1.0.0
